Normal Fixing
Seems the U2017 export is somehow stuffing my normal maps.  Not sure if a setting is wrong or if that's a new "feature".  Anyway, overwriting it directly from gimp works fine.
I also removed the spotlight items for now until I can tie their activity to the light action.:wq